# backend/openapi.yml
openapi: 3.0.3
info:
  title: STGY API
  version: 1.0.0
  description: |
    STGY SNS backend API (Express/TypeScript)
    - Cookie authentication: session_id
servers:
  - url: http://localhost:3001
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session_id
  schemas:
    ErrorResponse:
      type: object
      properties:
        error: { type: string }
      required: [error]
...
    GeoAddress:
      type: object
      properties:
        locale: { type: string }
        label: { type: string }
        elements:
          type: array
          items: { type: string }
      required: [locale, label, elements]
    GeoPlace:
      type: object
      properties:
        level: { type: integer, minimum: 1 }
        kind:
          type: string
          enum: [country, prefecture, municipality, special-ward, designated-city-ward]
        country: { type: string }
        longitude: { type: number, format: double }
        latitude: { type: number, format: double }
        addresses:
          type: array
          items:
            $ref: "#/components/schemas/GeoAddress"
      required: [level, kind, country, longitude, latitude, addresses]
    GeoPlaces:
      type: array
      items:
        $ref: "#/components/schemas/GeoPlace"
...
  responses:
    Error400:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Error401:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Error403:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Error404:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
paths:
...
  /geo/encode:
    get:
      summary: Encode a place name
      parameters:
        - in: query
          name: query
          required: true
          schema: { type: string }
        - in: query
          name: locale
          required: false
          description: Locale preference; exact locale, then base language, then the configured default are tried.
          schema: { type: string, default: en }
      responses:
        "200":
          description: Matching place and its parent hierarchy, ordered by descending level
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeoPlaces"
        "400":
          $ref: "#/components/responses/Error400"
        "401":
          $ref: "#/components/responses/Error401"
        "403":
          $ref: "#/components/responses/Error403"
        "404":
          $ref: "#/components/responses/Error404"
  /geo/decode:
    get:
      summary: Decode coordinates to a place
      parameters:
        - in: query
          name: longitude
          required: true
          schema: { type: number, format: double, minimum: -180, maximum: 180 }
        - in: query
          name: latitude
          required: true
          schema: { type: number, format: double, minimum: -90, maximum: 90 }
        - in: query
          name: locale
          required: false
          description: Locale preference; exact locale, then base language, then the configured default are tried.
          schema: { type: string, default: en }
      responses:
        "200":
          description: Nearest highest-level place and its parent hierarchy, ordered by descending level
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeoPlaces"
        "400":
          $ref: "#/components/responses/Error400"
        "401":
          $ref: "#/components/responses/Error401"
        "403":
          $ref: "#/components/responses/Error403"
        "404":
          $ref: "#/components/responses/Error404"
...